implemented api and test cases

+++ tests/conftest.py
@@ -1,4 +1,3 @@
-import logging
 import os
 import sys
 from datetime import datetime, timezone
@@ -10,67 +9,82 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import Session, sessionmaker
 from sqlalchemy_utils import create_database, database_exists, drop_database
 
-from casa.models import Account, Transaction  # Adjust the import according to your project structure
-
-logger = logging.getLogger(__name__)
-
-
 cwd = os.path.dirname(os.path.abspath(__file__))
 sys.path.insert(0, os.path.abspath(f"{cwd}/.."))
 
+from casa import models  # noqa
+from casa.api import db_session  # noqa
+from main import app  # noqa
 
-@pytest.fixture(scope="session")
-def sql_engine(tmp_path_factory):
-    test_db_url = os.environ.get("TEST_DATABASE_URL")
-    if test_db_url is None:
-        test_db_url = f"sqlite:///{tmp_path_factory.getbasetemp()}/test.db"
 
-    if test_db_url.startswith("postgres"):
-        test_db_url = test_db_url.set(
-            username=os.environ.get("TEST_PGUSER"),
-            password=os.environ.get("TEST_PGPASSWORD"),
-        )
+def tmp_sqlite_url():
+    tmp_path = os.path.abspath(f"{cwd}/../tmp")
+    os.makedirs(tmp_path, exist_ok=True)
+    return f"sqlite:///{tmp_path}/test.db"
 
+
+test_db_url = os.environ.get("TEST_DATABASE_URI", tmp_sqlite_url())
+testing_sql_engine = create_engine(test_db_url, connect_args={"check_same_thread": False})
+TestingSessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=testing_sql_engine)
+
+
+# Testing Dependency
+def testing_db_session():
+    session = TestingSessionLocal()
+    try:
+        yield session
+    finally:
+        session.close()
+
+
+app.dependency_overrides[db_session] = testing_db_session
+
+
+@pytest.fixture(autouse=True, scope="session")
+def test_db():
+    """
+    prepare the test database:
+    1. if the database does not exist, create it
+    2. run migration on the database
+    3. delete the database after the test, unless KEEP_TEST_DB is set to Y
+    """
+
+    test_db_created = False
     if not database_exists(test_db_url):
-        logger.info(f"creating test database {test_db_url}")
+        print(f"==creating test database {test_db_url}")
         create_database(test_db_url)
+        test_db_created = True
 
-    # Run the migrations
-    # migrations/env.py sets the sqlalchemy.url using env var DATABASE_URL
-    # so we set it so that alembic knows to use the correct database during testing
-    os.environ["DATABASE_URL"] = test_db_url
-    alembic_cfg = Config("alembic.ini")
-    command.upgrade(alembic_cfg, "head")
+        # Run the migrations
+        # so we set it so that alembic knows to use the correct database during testing
+        os.environ["SQLALCHEMY_DATABASE_URI"] = test_db_url
+        alembic_cfg = Config("alembic.ini")
+        command.upgrade(alembic_cfg, "head")
 
-    sql_engine = create_engine(test_db_url)
+        # seed test data
+        with TestingSessionLocal() as session:
+            print("==adding seed data")
+            seed_data(session)
 
-    # seed test data
-    with sessionmaker(bind=sql_engine)() as session:
-        seed_data(session)
+    yield testing_sql_engine  # the value is not used
 
-    yield sql_engine
-
-    if os.environ.get("KEEP_TEST_DB", "N").upper() not in ["1", "Y", "YES", "TRUE"]:
+    if test_db_created and os.environ.get("KEEP_TEST_DB", "N").upper() not in ["1", "Y", "YES", "TRUE"]:
         # drop the test database
-        logger.info(f"dropping test database {test_db_url}")
+        print(f"==dropping test database {test_db_url}")
         drop_database(test_db_url)
 
 
 @pytest.fixture(scope="session")
-def session(sql_engine):
-    Session = sessionmaker(bind=sql_engine)
-    session = Session()
-
-    yield session
-
-    session.close()
+def db_session():
+    with sessionmaker(autocommit=False, autoflush=False, bind=testing_sql_engine)() as session:
+        yield session
 
 
 def seed_data(session: Session):
     some_dt = datetime(2021, 1, 2, 12, 0, 1, tzinfo=timezone.utc)
 
     # create accounts
-    account1 = Account(
+    account1 = models.Account(
         account_num="1234567890",
         currency="USD",
         balance=1000.00,
@@ -78,7 +92,7 @@ def seed_data(session: Session):
         updated_at=some_dt,
     )
 
-    account2 = Account(
+    account2 = models.Account(
         account_num="0987654321",
         currency="USD",
         balance=500.00,
@@ -90,7 +104,7 @@ def seed_data(session: Session):
     session.commit()
 
     # create transactions
-    transaction1 = Transaction(
+    transaction1 = models.Transaction(
         ref_id="T1234567890",
         trx_date="2021-01-01",
         currency="USD",
@@ -100,7 +114,7 @@ def seed_data(session: Session):
         created_at=some_dt,
     )
 
-    transaction2 = Transaction(
+    transaction2 = models.Transaction(
         ref_id="T0987654321",
         trx_date="2021-01-01",
         currency="USD",
